requirements.txt
fastapi
uvicorn
pillow
python-dotenv
google-genai
aiohttp
python-multipart
pydantic